# swagger/swagger.json
{
  "swagger": "2.0",
  "paths": {
    "/fcc/config/list/": {
      "get": {
        "description": "Get config list",
        "tags": [
          "Fcc"
        ],
        "operationId": "FccConfigListReq",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ProjectKey",
            "name": "project_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "GroupKey",
            "name": "group_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ConfKey",
            "name": "conf_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchConfigsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/fcc/config/pre_publish/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Fcc"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FccPrePublishReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchGroupsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/fcc/config/publish/": {
      "post": {
        "description": "publish config",
        "tags": [
          "Fcc"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FccPublishReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/fcc/config/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Fcc"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FccSaveConfigReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/fcc/config/value/": {
      "get": {
        "description": "Get project list",
        "tags": [
          "Fcc"
        ],
        "operationId": "FccGetConfigReq",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ProjectKey",
            "name": "project_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "GroupKey",
            "name": "group_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ConfKey",
            "name": "conf_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/Config"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/fcc/group/list/": {
      "get": {
        "description": "Get group list",
        "tags": [
          "Fcc"
        ],
        "operationId": "FccGroupListReq",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ProjectKey",
            "name": "project_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "GroupKey",
            "name": "group_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "GroupName",
            "name": "group_name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchGroupsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/fcc/group/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Fcc"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FccSaveGroupReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/fcc/project/list/": {
      "get": {
        "description": "Get project list",
        "tags": [
          "Fcc"
        ],
        "operationId": "FccProjectListReq",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ProjectKey",
            "name": "project_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ProjectName",
            "name": "project_name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchProjectsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/fcc/project/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Fcc"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FccSaveProjectReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchGroupsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/starling/group/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Starling"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StarlingSaveGroupReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/starling/origin/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Starling"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StarlingSaveOriginLgReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/starling/project/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Starling"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StarlingSaveProjectReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchGroupsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/starling/translate/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Starling"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StarlingSaveTransLgReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "BaseResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/responses"
    },
...
    "Config": {
      "type": "object",
      "properties": {
        "conf_key": {
          "type": "string",
          "x-go-name": "ConfKey"
        },
        "create_time": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreateTime"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "op_id": {
          "type": "string",
          "x-go-name": "OpId"
        },
        "op_name": {
          "type": "string",
          "x-go-name": "OpName"
        },
        "pre_value": {
          "type": "string",
          "x-go-name": "PreValue"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "publish_status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "PublishStatus"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        },
        "update_time": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UpdateTime"
        },
        "value": {
          "type": "string",
          "x-go-name": "Value"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fcc_serv"
    },
    "FccPrePublishReq": {
      "type": "object",
      "properties": {
        "conf_key": {
          "type": "string",
          "x-go-name": "ConfKey"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "pre_value": {
          "type": "string",
          "x-go-name": "PreValue"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "FccPublishReq": {
      "type": "object",
      "properties": {
        "conf_key": {
          "type": "string",
          "x-go-name": "ConfKey"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "FccSaveConfigReq": {
      "type": "object",
      "properties": {
        "conf_key": {
          "type": "string",
          "x-go-name": "ConfKey"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "FccSaveGroupReq": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "group_name": {
          "type": "string",
          "x-go-name": "GroupName"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "FccSaveProjectReq": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "project_name": {
          "type": "string",
          "x-go-name": "ProjectName"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
...
    "FetchConfigsRet": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Config"
          },
          "x-go-name": "List"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fcc_serv"
    },
...
    "FetchGroupsRet": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Group"
          },
          "x-go-name": "List"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fly_starling_serv"
    },
...
    "FetchProjectsRet": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Project"
          },
          "x-go-name": "List"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fcc_serv"
    },
...
    "Group": {
      "type": "object",
      "properties": {
        "create_time": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreateTime"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "group_name": {
          "type": "string",
          "x-go-name": "GroupName"
        },
        "op_id": {
          "type": "string",
          "x-go-name": "OpId"
        },
        "op_name": {
          "type": "string",
          "x-go-name": "OpName"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        },
        "update_time": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UpdateTime"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fcc_serv"
    },
...
    "Project": {
      "type": "object",
      "properties": {
        "create_time": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreateTime"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "op_id": {
          "type": "string",
          "x-go-name": "OpId"
        },
        "op_name": {
          "type": "string",
          "x-go-name": "OpName"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "project_name": {
          "type": "string",
          "x-go-name": "ProjectName"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        },
        "update_time": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UpdateTime"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fly_starling_serv"
    },
...
    "StarlingSaveGroupReq": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "group_name": {
          "type": "string",
          "x-go-name": "GroupName"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "StarlingSaveOriginLgReq": {
      "type": "object",
      "properties": {
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "lang": {
          "type": "string",
          "x-go-name": "Lang"
        },
        "lang_key": {
          "type": "string",
          "x-go-name": "LangKey"
        },
        "origin_text": {
          "type": "string",
          "x-go-name": "OriginText"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "StarlingSaveProjectReq": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "project_name": {
          "type": "string",
          "x-go-name": "ProjectName"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "StarlingSaveTransLgReq": {
      "type": "object",
      "properties": {
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "lang": {
          "type": "string",
          "x-go-name": "Lang"
        },
        "lang_key": {
          "type": "string",
          "x-go-name": "LangKey"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        },
        "translate_text": {
          "type": "string",
          "x-go-name": "TranslateText"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
...
  }
}